requirements.txt
numpy>=1.21.0
scikit-learn>=0.24.0
scipy>=1.7.0
matplotlib>=3.4.0
streamlit>=1.0.0
llama-cpp-python
